8209345: Merge SATBMarkQueueFilter into SATBMarkQueueSet Summary: Move filter extension protocol to SATBMarkQueueSet. Reviewed-by: shade, tschatzl, rkennke
